requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
openai>=1.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
playwright>=1.40.0
pandas>=2.0.0
# Database dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 